Polish add button and css refactor on Layout folder
