+++ src/components/Layout/Footer.module.css
@@ -1,0 +1,23 @@
+.footer {
+  width: 100%;
+  padding: 22px 128px;
+  background: #091440;
+  bottom: 0;
+  position: fixed;
+}
+
+.textWrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+  justify-content: space-between;
+  height: 36px;
+  min-width: 210px;
+  font-size: 13px;
+  color: #8997a5;
+}
+
+.ghLink {
+  color: #2970ff;
+  text-decoration: none;
+}
