--- src/components/Layout/Layout.module.css
@@ -1,84 +1,0 @@
-/* Header styles */
-
-.headerContainer {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    height: 100%;
-    width: 100%;
-}
-
-.header {
-    color: #F9FAFA;
-    border: 0 !important;
-    height: 70px;
-}
-
-.spacerLeft {
-    padding-left: 20px; 
-}
-
-.spacerRight {
-    padding-right: 40px;
-}
-
-
-.searchBox {
-    width: 540px;
-    margin-top: 10px;
-}
-
-.searchInput {
-    border: 1px solid;
-    border-color: #A9B4BF;
-}
-
-.searchInput:hover {
-    border-color: rgba(41, 112, 255, 0.5);
-}
-
-.searchInput:focus {
-    border-color: #2970FF;
-}
-
-/* Page Body styles */
-
-.container {
-    background-color: #F3F5F6;
-}
-
-.pageMain {
-    max-width: 540px;
-    grid-row: auto !important;
-    grid-column: 5 / 14 !important;
-    margin-top: 6rem !important;
-    margin-bottom: 8rem !important;
-    margin-left: 50px !important;
-    padding: 0 0 0 0 !important;
-}
-
-/* Footer styles */
-
-.footer {
-    width: 100%;
-    padding: 22px 128px;
-    background: #091440;
-    bottom: 0;
-    position: fixed;
-}
-
-.textWrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    justify-content: space-between;
-    height: 36px;
-    min-width: 210px;
-    font-size: 13px;
-    color: #8997a5;
-}
-
-.ghLink {
-    color: #2970ff;
-    text-decoration: none;
-}
