+++ src/components/Layout/Page.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Header from "./Header";
 import Footer from "./Footer";
-import styles from "./Layout.module.css";
+import styles from "./Page.module.css";
 import { Container, Main } from "pi-ui";
 
 const Page = ({ children }) => {
